refactor: refactor file cjsToEsm.cjs - rewrite it to it's equivalent in ECMAScript notation
